Fix compiler warnings

+++ src/fcmatch.c
@@ -33,7 +33,7 @@ FcCompareNumber (FcValue *value1, FcValue *value2)
 {
     double  v1, v2, v;
 
-    switch (value1->type) {
+    switch ((int) value1->type) {
     case FcTypeInteger:
 	v1 = (double) value1->u.i;
 	break;
@@ -43,7 +43,7 @@ FcCompareNumber (FcValue *value1, FcValue *value2)
     default:
 	return -1.0;
     }
-    switch (value2->type) {
+    switch ((int) value2->type) {
     case FcTypeInteger:
 	v2 = (double) value2->u.i;
 	break;
@@ -86,9 +86,9 @@ FcCompareLang (FcValue *v1, FcValue *v2)
     FcLangResult    result;
     FcValue value1 = FcValueCanonicalize(v1), value2 = FcValueCanonicalize(v2);
 
-    switch (value1.type) {
+    switch ((int) value1.type) {
     case FcTypeLangSet:
-	switch (value2.type) {
+	switch ((int) value2.type) {
 	case FcTypeLangSet:
 	    result = FcLangSetCompare (value1.u.l, value2.u.l);
 	    break;
@@ -101,7 +101,7 @@ FcCompareLang (FcValue *v1, FcValue *v2)
 	}
 	break;
     case FcTypeString:
-	switch (value2.type) {
+	switch ((int) value2.type) {
 	case FcTypeLangSet:
 	    result = FcLangSetHasLang (value2.u.l, value1.u.s);
 	    break;
@@ -146,7 +146,7 @@ FcCompareSize (FcValue *value1, FcValue *value2)
 {
     double  v1, v2, v;
 
-    switch (value1->type) {
+    switch ((int) value1->type) {
     case FcTypeInteger:
 	v1 = value1->u.i;
 	break;
@@ -156,7 +156,7 @@ FcCompareSize (FcValue *value1, FcValue *value2)
     default:
 	return -1;
     }
-    switch (value2->type) {
+    switch ((int) value2->type) {
     case FcTypeInteger:
 	v2 = value2->u.i;
 	break;
